+++ airsen-backend/src/main/java/fr/airsen/api/dto/DepartmentDTO.java
@@ -1,22 +1,22 @@
 package fr.airsen.api.dto;
 
 public class DepartmentDTO {
-    private int id;
+    private Long id;
     private String name;
     private int departmentCode;
     private String regionCode;
 
     public DepartmentDTO() {}
 
-    public DepartmentDTO(int id, String name, int departmentCode, String regionCode) {
+    public DepartmentDTO(Long id, String name, int departmentCode, String regionCode) {
         this.id = id;
         this.name = name;
         this.departmentCode = departmentCode;
         this.regionCode = regionCode;
     }
 
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
