+++ airsen-backend/src/main/java/fr/airsen/api/dto/RegionDTO.java
@@ -1,20 +1,20 @@
 package fr.airsen.api.dto;
 
 public class RegionDTO {
-    private int id;
+    private Long id;
     private String name;
     private String regionCode;
 
     public RegionDTO() {}
 
-    public RegionDTO(int id, String name, String regionCode) {
+    public RegionDTO(Long id, String name, String regionCode) {
         this.id = id;
         this.name = name;
         this.regionCode = regionCode;
     }
 
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
