+++ airsen-backend/src/main/java/fr/airsen/api/repository/DepartmentRepository.java
@@ -5,8 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DepartmentRepository extends JpaRepository<Department, Integer> {
-
-    // Lister les départements par région
-    List<Department> findByRegionId(int regionId);
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
+    List<Department> findByRegionId(Long regionId);
 }
