+++ airsen-backend/src/main/java/fr/airsen/api/service/DepartmentService.java
@@ -21,7 +21,7 @@ public class DepartmentService {
     /**
      * Récupère un département pour un id donné.
      */
-    public DepartmentDTO getDepartmentById(int id) {
+    public DepartmentDTO getDepartmentById(Long id) {
         Optional<Department> opt = departmentRepository.findById(id);
         if (opt.isEmpty()) return null;
         Department d = opt.get();
@@ -36,7 +36,7 @@ public class DepartmentService {
     /**
      * Récupère tous les départements d'une région donnée.
      */
-    public List<DepartmentDTO> getDepartmentsByRegion(int regionId) {
+    public List<DepartmentDTO> getDepartmentsByRegion(Long regionId) {
         List<Department> departments = departmentRepository.findByRegionId(regionId);
         return departments.stream()
                 .map(d -> new DepartmentDTO(
